a2a/pyproject.toml
[project]
name = "a2a-agents"
dynamic = ["version"]
description = "A2A Granite Chat"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
 "agentstack-sdk>=0.4.2",
 "deepagents>=0.2.8",
 "granite-core",
 "langchain-ollama>=1.0.0",
 "langchain-openai>=1.0.2",
]

[tool.ruff]
lint.select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "UP",    # pyupgrade
    "I",     # isort
    "B",     # bugbear
    "ANN",   # annotations
    "N",     # pep8-naming
    "C4",    # Comprehensions
    "DTZ",   # DatetimeZ
    "Q",     # Quotes
    "SIM",   # Simplify
    "RUF",   # Ruff
    "TID",   # tidy-imports
    "ASYNC", # async
]
lint.ignore = ["ANN401"]
lint.fixable = ["ALL"]
lint.unfixable = []
force-exclude = true
line-length = 120
extend-exclude = []

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-tidy-imports]
# Disallow all relative imports, "parents" | "all"
ban-relative-imports = "all"

[tool.uv.sources]
detect-secrets = { git = "https://github.com/ibm/detect-secrets.git", rev = "master" }
granite-core = { path = "../granite_core", editable = true }

[dependency-groups]
dev = [
    "detect-secrets",
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.2",
    "ruff>=0.11.10",
]

[tool.pytest.ini_options]
addopts = "--maxfail=1"

[tool.setuptools]
packages = ["a2a_agents"]
